Make this work with both Xcode and Swift Build
Swift Build needs a separate target to get the C definitions imported,
so I use the dummy.m with the CSkiaSharp target, which I then import

I can not figure out how to make Xcode build the native headers into a
different module name.

Perhaps one way of solving this is to figure out a way of making the
Swift Build handle the C code in the same target as the Swift code.